requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
elevenlabs>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
asyncpg>=0.28.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
